Refactor Header component to separate the server and client component
