+++ web/src/components/common/Header.module.css
@@ -28,156 +28,8 @@
     transition: opacity 0.2s ease;
   }
   
-  /* Desktop Navigation */
-  .desktopNav {
-    display: flex;
-    align-items: center;
-    gap: 2rem;
-  }
-  
-  .navList {
-    display: flex;
-    list-style: none;
-    gap: 2rem;
-    margin: 0;
-    padding: 0;
-  }
-  
-  .navLink {
-    color: var(--primary-text-color);
-    font-weight: 500;
-    transition: color 0.2s ease;
-    padding: 0.5rem 0;
-  }
-  
-  .navLink:hover {
-    color: var(--primary-color);
-  }
-  
-  .loginButton {
-    background-color: var(--primary-color);
-    color: var(--dark-background);
-    border: none;
-    padding: 0.75rem 1.5rem;
-    border-radius: 5px;
-    font-weight: 600;
-    cursor: pointer;
-    transition: background-color 0.2s ease;
-  }
-  
-  .loginButton:hover {
-    background-color: #e68a00;
-  }
-  
-  /* Mobile Menu Button */
-  .mobileMenuButton {
-    display: none;
-    background: none;
-    border: none;
-    cursor: pointer;
-    padding: 0.5rem;
-  }
-  
-  .hamburger {
-    display: flex;
-    flex-direction: column;
-    width: 24px;
-    height: 18px;
-    position: relative;
-  }
-  
-  .hamburger span {
-    display: block;
-    height: 2px;
-    width: 100%;
-    background-color: var(--primary-text-color);
-    margin-bottom: 4px;
-    transition: all 0.3s ease;
-    transform-origin: center;
-  }
-  
-  .hamburger span:last-child {
-    margin-bottom: 0;
-  }
-  
-  .hamburger.open span:nth-child(1) {
-    transform: rotate(45deg) translate(6px, 6px);
-  }
-  
-  .hamburger.open span:nth-child(2) {
-    opacity: 0;
-  }
-  
-  .hamburger.open span:nth-child(3) {
-    transform: rotate(-45deg) translate(6px, -6px);
-  }
-  
-  /* Mobile Navigation */
-  .mobileNav {
-    display: none;
-    position: absolute;
-    top: 100%;
-    left: 0;
-    right: 0;
-    background-color: var(--dark-background);
-    border-top: 1px solid rgba(255, 255, 255, 0.1);
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  }
-  
-  .mobileNav.open {
-    display: block;
-  }
-  
-  .mobileNavList {
-    list-style: none;
-    padding: 1rem;
-    margin: 0;
-  }
-  
-  .mobileNavList li {
-    margin-bottom: 0.5rem;
-  }
-  
-  .mobileNavLink {
-    display: block;
-    color: var(--primary-text-color);
-    padding: 0.75rem 0;
-    font-weight: 500;
-    border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-    transition: color 0.2s ease;
-  }
-  
-  .mobileNavLink:hover {
-    color: var(--primary-color);
-  }
-  
-  .mobileLoginButton {
-    width: 100%;
-    background-color: var(--primary-color);
-    color: var(--dark-background);
-    border: none;
-    padding: 0.75rem;
-    border-radius: 5px;
-    font-weight: 600;
-    cursor: pointer;
-    margin-top: 1rem;
-    transition: background-color 0.2s ease;
-  }
-  
-  .mobileLoginButton:hover {
-    background-color: #e68a00;
-  }
-  
   /* Responsive Design */
   @media (max-width: 768px) {
-    .desktopNav {
-      display: none;
-    }
-  
-    .mobileMenuButton {
-      display: block;
-    }
-  
     .container {
       padding: 0 1rem;
     }
@@ -185,6 +37,7 @@
     .logo {
       font-size: 1.25rem;
     }
+
   }
   
   @media (max-width: 480px) {
@@ -197,86 +50,3 @@
       font-size: 1.125rem;
     }
   }
-  
-/* User Auth Design */ 
-.authSection {
-  display: flex;
-  align-items: center;
-}
-
-.mobileAuthSection {
-  margin-top: 16px;
-  padding-top: 16px;
-  border-top: 1px solid #e5e7eb;
-}
-
-.mobileUserSection {
-  padding: 16px 0;
-}
-
-.mobileUserInfo {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  margin-bottom: 16px;
-  padding: 0 16px;
-}
-
-.mobileUserName {
-  font-weight: 600;
-  font-size: 16px;
-}
-
-.mobileUserRole {
-  font-size: 14px;
-  text-transform: capitalize;
-}
-
-.mobileUserLinks {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.mobileLogoutButton {
-  display: block;
-  width: 100%;
-  padding: 12px 16px;
-  text-align: left;
-  background: none;
-  border: none;
-  color: #dc2626;
-  font-size: 16px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: background-color 0.2s ease;
-}
-
-.mobileLogoutButton:hover {
-  background: #fef2f2;
-}
-
-.loadingSpinner {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 8px;
-}
-
-.spinner {
-  width: 20px;
-  height: 20px;
-  border: 2px solid #e5e7eb;
-  border-top: 2px solid #3b82f6;
-  border-radius: 50%;
-  animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
