+++ web/src/components/common/header/MobileNav.module.css
@@ -1,0 +1,131 @@
+.mobileNav {
+  display: none;
+  position: absolute;
+  top: 100%;
+  left: 0;
+  right: 0;
+  background-color: var(--dark-background);
+  border-top: 1px solid rgba(255, 255, 255, 0.1);
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+.mobileNav.open {
+  display: block;
+}
+
+.mobileNavList {
+  list-style: none;
+  padding: 1rem;
+  margin: 0;
+}
+
+.mobileNavList li {
+  margin-bottom: 0.5rem;
+}
+
+.mobileNavLink {
+  display: block;
+  color: var(--primary-text-color);
+  padding: 0.75rem 0;
+  font-weight: 500;
+  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+  transition: color 0.2s ease;
+}
+
+.mobileNavLink:hover {
+  color: var(--primary-color);
+}
+
+.mobileLoginButton {
+  width: 100%;
+  background-color: var(--primary-color);
+  color: var(--dark-background);
+  border: none;
+  padding: 0.75rem;
+  border-radius: 5px;
+  font-weight: 600;
+  cursor: pointer;
+  margin-top: 1rem;
+  transition: background-color 0.2s ease;
+}
+
+.mobileLoginButton:hover {
+  background-color: #e68a00;
+}
+
+.mobileAuthSection {
+  margin-top: 16px;
+  padding-top: 16px;
+  border-top: 1px solid #e5e7eb;
+}
+
+.mobileUserSection {
+  padding: 16px 0;
+}
+
+.mobileUserInfo {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  margin-bottom: 16px;
+  padding: 0 16px;
+}
+
+.mobileUserName {
+  font-weight: 600;
+  font-size: 16px;
+}
+
+.mobileUserRole {
+  font-size: 14px;
+  text-transform: capitalize;
+}
+
+.mobileUserLinks {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.mobileLogoutButton {
+  display: block;
+  width: 100%;
+  padding: 12px 16px;
+  text-align: left;
+  background: none;
+  border: none;
+  color: #dc2626;
+  font-size: 16px;
+  font-weight: 500;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+}
+
+.mobileLogoutButton:hover {
+  background: #fef2f2;
+}
+
+.loadingSpinner {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 8px;
+}
+
+.spinner {
+  width: 20px;
+  height: 20px;
+  border: 2px solid #e5e7eb;
+  border-top: 2px solid #3b82f6;
+  border-radius: 50%;
+  animation: spin 1s linear infinite;
+}
+
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
